pyproject.toml
[tool.poetry]
name = "m3l"
version = "0.0.0" 
description = "MultiMedia and Machine Learning"
authors = [
    "Max Ehrlich"
]

license = "MIT"
readme = "README.md"

homepage = ""
repository = ""
documentation = ""


classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 3 - Alpha"
]

[tool.poetry.dependencies]
python = ">= 3.7, < 4.0"
pillow = "^8.0.1"

torch = "1.8.1"
torchvision = "0.9.1"
torchjpeg = "0.9.17"
torchtyping = "^0.1.1"
pytorch-lightning = "1.3.3"

hydra-core = "1.1.0.rc1"
hydra-colorlog = "1.1.0.dev1"

submitit = { version = "^1.3.3", optional = true }
hydra-submitit-launcher = { version = "^1.1.5.dev2", optional = true }

comet-ml = { version = "^3.7.0", optional = true }
wandb = { version = "^0.10.31", optional = true }
neptune-client = { version = "^0.9.15", optional = true }

bitstring = "^3.1.7"
av = "^8.0.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
pylint = "^2.6.0"

[tool.poetry.extras]
slurm = ["submitit", "hydra-submitit-launcher"]

comet-ml = ["comet-ml"]
wandb = ["wandb"]
neptune = ["neptine-client"]

[build-system]
requires = [
    "poetry>=1.0.10",
]
build-backend = "poetry.masonry.api"